requirements.txt
absl-py==2.1.0
accelerate==0.30.1
chex==0.1.85
evaluate==0.4.1
langdetect==1.0.9
ml-collections==0.1.1
numpy==1.26.4
omegaconf==2.3.0
peft==0.11.1
protobuf==5.26.1
Pygments==2.17.2
sacrebleu==2.4.2
scikit-learn==1.4.2
scipy==1.12.0
sentencepiece==0.2.0
transformers==4.41.0
